Implements liking posts via Redux
Refactors the like functionality to use Redux for state management.

This change introduces a new Redux action, `likePost`, which dispatches an optimistic update to the store before attempting to update the server. Upon successful server update, the store is updated with the correct data. In case of failure, a notification is dispatched.

Additionally, sorts the blogs by likes on the client side to avoid sorting bugs.